requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.4.0
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.5
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.21.3
opencv-python==4.5.4.58
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
portpicker==1.2.0
prometheus-client==0.13.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
requests==2.21.0
rsa==4.8
Send2Trash==1.8.0
simplegeneric==0.8.1
six==1.12.0
sniffio==1.2.0
stack-data==0.2.0
terminado==0.13.1
testpath==0.6.0
traitlets==5.1.1
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
